reuse thinning levels config between apps
