+++ libs/api-client/src/index.ts
@@ -2,3 +2,4 @@ export * from './api-client'
 export * from './utils/url'
 export * from './utils/search'
 export * from './utils/errors'
+export * from './utils/thinning'
